Automation to dim the screen after idle timeout
Cribbed from the change-page-on-idle automation, this one alters the
screen brightness (using MQTT, not the Home-Assistant Concept of
Backlight).
the value of "dim" and "bright" can be set, and so can the timeout.

Drawbacks: now every button_short_press causes a brightness change
command to be sent. And this doesn't honour the Home-Assistant
Concept of Backlight slider, which is not intuitive and could be
done better.

Do sliders generate button_short_press events?