Add option to drop touch events when backlight is off
This adds functionality to optionally drop touch events when the
backlight is off and send an mqtt message. This allows for a child lock
mode or the backlight to turn off on a schedule but turn back on when
the user touches the screen without it sending an event until the
backlight is on again.